+++ src/map/status.h
@@ -53,10 +53,10 @@ short status_get_clothes_color(struct block_list *bl);
 
 struct status_change *status_get_sc_data(struct block_list *bl);
 short *status_get_sc_count(struct block_list *bl);
-short *status_get_opt1(struct block_list *bl);
-short *status_get_opt2(struct block_list *bl);
-short *status_get_opt3(struct block_list *bl);
-short *status_get_option(struct block_list *bl);
+unsigned short *status_get_opt1(struct block_list *bl);
+unsigned short *status_get_opt2(struct block_list *bl);
+unsigned int *status_get_opt3(struct block_list *bl);
+unsigned int *status_get_option(struct block_list *bl);
 
 int status_get_matk1(struct block_list *bl);
 int status_get_matk2(struct block_list *bl);
